Convert number codes to string for VSCode's IMarkerData
VSCode's IMarkerData expects its code property to be a string. The
LSP's Diagnostic code however can be a string or a number. Casting
the value to a string in TypeScript will not actually transform the
value into a string so we should manually call toString() on the
value if it is actually a number.

Signed-off-by: Remy Suen <remy.suen@gmail.com>