Create a browser-only example
As the Monaco Editor can be opened in a simple HTML page, it should
also be possible to leverage the monaco-languageclient in the same
setup. TO demonstrate this, a new example has been created in the
examples/browser folder.

Instead of using the actual language server for the editor features,
the language service that is backing the language server must be used
instead. The types from the language service will be converted to and
from Monaco types with the help of the converters provided by
monaco-languageclient.

Since this is intended to be run in a simple HTML page without a
server serving the content, only language services written in
JavaScript may be used as that is the programming language that
browsers speak.

Signed-off-by: Remy Suen <remy.suen@gmail.com>