Close websocket
When I implemented the current react-client example on a long-running server instance, I saw exploding memory usage but no CPU usage. The memory hung for a long time but was never released. I checked my console network tab, and the connections were never being closed. 
I believe there need to be calls made to close the web sockets, or else there will be a build up of memory on the server.